Add files via upload
I've made this project using machine learning. Nowadays we have received so many spam mails, so for this reason we have to sort out which are important and which are spam. So, at first I train the model, then test this. I get approximately same accuracy for train and test dataset. And my project is very well to detect spam mails.